requirements.txt
Django>=3.2
nepali-datetime>=1.0.0
pytz>=2021.1
python-dateutil>=2.8
djangorestframework>=3.12
pytest>=7.0
pytest-django>=4.5